Adding new update sermon summaries script for summary migration
This script is useful to create the summaries once at first should none exist. And for prompt changes. Should your prompts start generating bad results, this tool can be used in conjunction with the agents modified prompt to update all messages in bulk and save time when migrating to a new prompt.